9_vector_databases/7_milvus/requirements.txt
langchain
langchain-core
langchain-openai
langchain-milvus
pymilvus
python-dotenv
numpy
rich